auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
sqlalchemy==2.0.23
httpx==0.25.1
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9